refactor: organized applications and applications details page components
