+++ jsconfig.json
@@ -1,10 +1,8 @@
 {
 	"compilerOptions": {
-		// ...
 		"baseUrl": ".",
 		"paths": {
 			"@/*": ["./src/*"]
 		}
-		// ...
 	}
 }
