+++ src/components/applications/ApplicationsHeading.jsx
@@ -1,8 +1,13 @@
-import React from "react";
-import ArrowDown from "../../../assets/icons/ArrowDown";
-import SearchIcon from "../../../assets/icons/SearchIcon";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import SearchIcon from "../../assets/icons/SearchIcon";
 
-const ApplicationHeading = () => {
+const ApplicationsHeading = () => {
   return (
     <div className="flex items-center justify-between mb-[24px] p-[16px_23px_16px_17px] border-b-[1px] border-[#E5E5E5]">
       <h3 className="text-[#111928] text-[18px] leading-[28px] font-semibold">
@@ -22,12 +27,19 @@ const ApplicationHeading = () => {
             <SearchIcon color="#046C4E" />
           </div>
         </div>
-        <button className="flex items-center gap-x-[10px] text-[#6B7280] bg-[#F9FAFB] p-[12px_16px] border-[1px] border-[#D1D5DB] rounded-lg">
-          Sort By: New <ArrowDown />
-        </button>
+        <Select>
+          <SelectTrigger className="w-[180px]">
+            <SelectValue placeholder="Sort By: New" />
+          </SelectTrigger>
+          <SelectContent>
+            <SelectItem value="light">New</SelectItem>
+            <SelectItem value="dark">Dark</SelectItem>
+            <SelectItem value="system">System</SelectItem>
+          </SelectContent>
+        </Select>
       </div>
     </div>
   );
 };
 
-export default ApplicationHeading;
+export default ApplicationsHeading;
