+++ src/components/applications/applications-details/ApplicantPhoto.jsx
