+++ src/components/applications/applications-details/ApplicationDetailsTabs.jsx
